% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionHelpers.R
\name{cvperformance}
\alias{cvperformance}
\title{cvperformance}
\usage{
cvperformance(data, model)
}
\arguments{
\item{data}{un dataframe contenant la variable réponse ModeAccouc et les
variables explicatives numériques sélctionnées par l'utilisateur}

\item{model}{le type de modèle choisi dans c("multinom", "knn")}
}
\value{
acc : accuracy du modèle calculée par validation croisée en 10
segments
k : hyperparamètre du modèle knn (si model="knn")
}
\description{
Cette fonction évalue la performance de prédiction de deux modèles (multinom
et knn) en validation croisée par 10 segments
}
