% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVisServer.R
\name{DataVisServer}
\alias{DataVisServer}
\title{DataVisServer}
\usage{
DataVisServer(input, output)
}
\arguments{
\item{output}{}
}
\value{
output
}
\description{
Creates and exports the server for the 'data visualisation' part of the app
}
