% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionHelpers.R
\name{get.newdata}
\alias{get.newdata}
\title{get.newdata}
\usage{
get.newdata(var, ipt)
}
\arguments{
\item{var}{variable à tester, comprise dans le vecteur values :
c("Nbsem","Agedelamere","TailMere","PoidsMere","Agedupere","TailPere",
"PoidsPere","NbGrossess","NbEnfants","NbIVG","NbFC")}

\item{ipt}{la reactiveValues input est transformée en liste par
reactiveValuesToList()}
}
\value{
la fonction renvoie la valeur entrée par l'utilisateur, en vérifiant
au préalable que les données entrées sont correctes (format numérique et
comprises dans les variations normales de la variable avec une marge de 20%)
}
\description{
get.newdata
}
