% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launchApp}
\usage{
launchApp(launch.browser = T)
}
\arguments{
\item{launch.browser}{boolean value indicating whether the app should be
launched in the user's default browser (T), or in a R window (F)}
}
\description{
Launches the shiny app
}
\examples{
launchApp() #in the browser
launchApp(launch.browser=F) #in a R window

}
