% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModellingServer.R
\name{ModellingServer}
\alias{ModellingServer}
\title{ModellingServer}
\usage{
ModellingServer(input, output)
}
\arguments{
\item{output}{}
}
\value{
output
}
\description{
Creates and exports the server for the 'modelling' part of the app
}
