% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionServer.R
\name{PredictionServer}
\alias{PredictionServer}
\title{PredictionServer}
\usage{
PredictionServer(input, output)
}
\arguments{
\item{output}{}
}
\value{
output
}
\description{
Creates and exports the server for the 'prediction' part of the app
}
